--- wls-wahlvorbereitung-service/src/main/java/de/muenchen/oss/wahllokalsystem/wahlvorbereitung/configuration/ForwardedHeaderConfiguration.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
- * der Landeshauptstadt München, 2024
- */
-package de.muenchen.oss.wahllokalsystem.wahlvorbereitung.configuration;
-
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.filter.ForwardedHeaderFilter;
-
-/**
- * This class provides the {@link ForwardedHeaderFilter} to handle
- * the headers of type "Forwarded" and "X-Forwarded-*".
- */
-@Configuration
-public class ForwardedHeaderConfiguration {
-
-    @Bean
-    public FilterRegistrationBean<ForwardedHeaderFilter> forwardedHeaderFilter() {
-        final FilterRegistrationBean<ForwardedHeaderFilter> registration = new FilterRegistrationBean<>();
-        registration.setFilter(new ForwardedHeaderFilter());
-        registration.addUrlPatterns("/*");
-        return registration;
-    }
-
-}
