--- wls-wahlvorbereitung-service/src/main/java/de/muenchen/oss/wahllokalsystem/wahlvorbereitung/domain/TheEntity.java
@@ -1,39 +1,0 @@
-package de.muenchen.oss.wahllokalsystem.wahlvorbereitung.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-/**
- * This class represents a TheEntity.
- * <p>
- * The entity's content will be loaded according to the reference variable.
- * </p>
- */
-@Entity
-// Definition of getter, setter, ...
-@Getter
-@Setter
-@ToString(callSuper = true)
-@EqualsAndHashCode(callSuper = true)
-@NoArgsConstructor
-public class TheEntity extends BaseEntity {
-
-    private static final long serialVersionUID = 1L;
-
-    // ========= //
-    // Variables //
-    // ========= //
-
-    @Column(name = "textattribute", nullable = false, length = 8)
-    @NotNull
-    @Size(min = 2, max = 8)
-    private String textAttribute;
-
-}
