+++ wls-wahlvorbereitung-service/src/main/java/de/muenchen/oss/wahllokalsystem/wahlvorbereitungservice/configuration/SwaggerConfiguration.java
@@ -1,0 +1,43 @@
+package de.muenchen.oss.wahllokalsystem.wahlvorbereitungservice.configuration;
+
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Contact;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.security.SecurityRequirement;
+import io.swagger.v3.oas.models.security.SecurityScheme;
+import org.springdoc.core.models.GroupedOpenApi;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfiguration {
+
+    @Value("${info.application.version:unknown}")
+    String version;
+
+    @Bean
+    GroupedOpenApi publicApi() {
+        return GroupedOpenApi.builder()
+                .group("public-apis")
+                .pathsToMatch("/**")
+                .build();
+    }
+
+    @Bean
+    OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .info(
+                        new Info().title("Wahlvorbereitung Service")
+                                .version(version)
+                                .contact(new Contact().name("Your Name").email("Your E-Mail-Address")))
+                .addSecurityItem(new SecurityRequirement().addList("bearerAuth"))
+                .components(
+                        new Components()
+                                .addSecuritySchemes("bearerAuth", new SecurityScheme()
+                                        .type(SecurityScheme.Type.HTTP)
+                                        .scheme("bearer")
+                                        .bearerFormat("JWT")));
+    }
+}
