--- wls-wahlvorbereitung-service/sso-config/sso-authorisation.json
@@ -1,72 +1,0 @@
-{
-    "allowRemoteResourceManagement": false,
-    "policyEnforcementMode": "ENFORCING",
-    "resources": [
-        {
-            "name": "WLS_WAHLVORBEREITUNG_SERVICE_READ_THEENTITY",
-            "type": "wahlvorbereitung-wls-wahlvorbereitung-service-service",
-            "scopes": []
-        },
-        {
-            "name": "WLS_WAHLVORBEREITUNG_SERVICE_WRITE_THEENTITY",
-            "type": "wahlvorbereitung-wls-wahlvorbereitung-service-service",
-            "scopes": []
-        },
-        {
-            "name": "WLS_WAHLVORBEREITUNG_SERVICE_DELETE_THEENTITY",
-            "type": "wahlvorbereitung-wls-wahlvorbereitung-service-service",
-            "scopes": []
-        }
-    ],
-    "policies": [
-        {
-            "name": "rolepolicy_admin",
-            "type": "role",
-            "logic": "POSITIVE",
-            "decisionStrategy": "AFFIRMATIVE",
-            "config": {
-                "roles": "[{\"id\":\"wahlvorbereitung/clientrole_admin\",\"required\":false}]"
-            }
-        },
-        {
-            "name": "rolepolicy_readonly",
-            "type": "role",
-            "logic": "POSITIVE",
-            "decisionStrategy": "AFFIRMATIVE",
-            "config": {
-                "roles": "[{\"id\":\"wahlvorbereitung/clientrole_readonly\",\"required\":false}]"
-            }
-        },
-        {
-            "name": "permission_WLS_WAHLVORBEREITUNG_SERVICE_DELETE_THEENTITY",
-            "type": "resource",
-            "logic": "POSITIVE",
-            "decisionStrategy": "AFFIRMATIVE",
-            "config": {
-                "resources": "[\"WLS_WAHLVORBEREITUNG_SERVICE_DELETE_THEENTITY\"]",
-                "applyPolicies": "[\"rolepolicy_admin\"]"
-            }
-        },
-        {
-            "name": "permission_WLS_WAHLVORBEREITUNG_SERVICE_WRITE_THEENTITY",
-            "type": "resource",
-            "logic": "POSITIVE",
-            "decisionStrategy": "AFFIRMATIVE",
-            "config": {
-                "resources": "[\"WLS_WAHLVORBEREITUNG_SERVICE_WRITE_THEENTITY\"]",
-                "applyPolicies": "[\"rolepolicy_admin\"]"
-            }
-        },
-        {
-            "name": "permission_WLS_WAHLVORBEREITUNG_SERVICE_READ_THEENTITY",
-            "type": "resource",
-            "logic": "POSITIVE",
-            "decisionStrategy": "AFFIRMATIVE",
-            "config": {
-                "resources": "[\"WLS_WAHLVORBEREITUNG_SERVICE_READ_THEENTITY\"]",
-                "applyPolicies": "[\"rolepolicy_admin\",\"rolepolicy_readonly\"]"
-            }
-        }
-    ],
-    "scopes": []
-}
