--- wls-wahlvorbereitung-service/sso-config/sso-client.json
@@ -1,70 +1,0 @@
-{
-  "clientId": "wahlvorbereitung",
-  "surrogateAuthRequired": false,
-  "enabled": true,
-  "clientAuthenticatorType": "client-secret",
-  "redirectUris": [ "*" ],
-  "webOrigins": [],
-  "notBefore": 0,
-  "bearerOnly": false,
-  "consentRequired": false,
-  "standardFlowEnabled": true,
-  "implicitFlowEnabled": false,
-  "directAccessGrantsEnabled": true,
-  "serviceAccountsEnabled": true,
-  "authorizationServicesEnabled": true,
-  "publicClient": false,
-  "frontchannelLogout": false,
-  "protocol": "openid-connect",
-  "attributes": {
-    "saml.assertion.signature": "false",
-    "saml.force.post.binding": "false",
-    "saml.multivalued.roles": "false",
-    "saml.encrypt": "false",
-    "saml_force_name_id_format": "false",
-    "saml.client.signature": "false",
-    "saml.authnstatement": "false",
-    "saml.server.signature": "false",
-    "saml.server.signature.keyinfo.ext": "false",
-    "saml.onetimeuse.condition": "false"
-  },
-  "fullScopeAllowed": true,
-  "nodeReRegistrationTimeout": -1,
-  "protocolMappers": [
-    {
-      "name": "Client ROLES to user_role",
-      "protocol": "openid-connect",
-      "protocolMapper": "oidc-usermodel-client-role-mapper",
-      "consentRequired": false,
-      "consentText": "",
-      "config": {
-        "usermodel.clientRoleMapping.rolePrefix": "ROLE_",
-        "multivalued": "true",
-        "userinfo.token.claim": "true",
-        "id.token.claim": "true",
-        "access.token.claim": "true",
-        "claim.name": "user_roles",
-        "jsonType.label": "String",
-        "usermodel.clientRoleMapping.clientId": "wahlvorbereitung"
-      }
-    },
-    {
-      "name": "Client PERMISSION to Spring Security authorities",
-      "protocol": "openid-connect",
-      "protocolMapper": "oidc-authorities-mapper",
-      "consentRequired": false,
-      "config": {
-        "authorities": "authorities",
-        "userinfo.token.claim": "true"
-      }
-    }
-  ],
-  "useTemplateConfig": false,
-  "useTemplateScope": false,
-  "useTemplateMappers": false,
-  "access": {
-    "view": true,
-    "configure": true,
-    "manage": true
-  }
-}
